requirements.txt
pygame
pytz
loguru
schedule
pytest
todoist-python
pillow
dotmap
python-dotenv